Enable REST api (using expressjs)
