+++ config-overrides.js
@@ -1,0 +1,5 @@
+/* config-overrides.js */
+/* eslint-disable react-hooks/rules-of-hooks */
+const { useBabelRc, override } = require('customize-cra');
+
+module.exports = override(useBabelRc());
